backend-ai/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
google-genai>=1.0.0
pydantic>=2.0.0
python-dotenv==1.0.1
httpx>=0.24.0